+++ src/Enums/SharpApiJobTypeEnum.php
@@ -27,7 +27,35 @@ enum SharpApiJobTypeEnum: string
     case CONTENT_SUMMARIZE = 'content_summarize';
     case CONTENT_KEYWORDS = 'content_keywords';
     case CONTENT_TRANSLATE = 'content_translate';
+    case CONTENT_PARAPHRASE = 'content_paraphrase';
+    case CONTENT_PROOFREAD = 'content_proofread';
     case SEO_GENERATE_TAGS = 'seo_generate_tags';
+
+    public function url(): string
+    {
+        return match ($this) {
+            self::ECOMMERCE_REVIEW_SENTIMENT => '/ecommerce/review_sentiment',
+            self::ECOMMERCE_PRODUCT_CATEGORIES => '/ecommerce/product_categories',
+            self::ECOMMERCE_PRODUCT_INTRO => '/ecommerce/product_intro',
+            self::ECOMMERCE_THANK_YOU_EMAIL => '/ecommerce/thank_you_email',
+            self::HR_PARSE_RESUME => '/hr/parse_resume',
+            self::HR_JOB_DESCRIPTION => '/hr/job_description',
+            self::HR_RELATED_SKILLS => '/hr/related_skills',
+            self::HR_RELATED_JOB_POSITIONS => '/hr/related_job_positions',
+            self::TTH_REVIEW_SENTIMENT => '/tth/review_sentiment',
+            self::TTH_TA_PRODUCT_CATEGORIES => '/tth/ta_product_categories',
+            self::TTH_HOSPITALITY_PRODUCT_CATEGORIES => '/tth/hospitality_product_categories',
+            self::CONTENT_DETECT_PHONES => '/content/detect_phones',
+            self::CONTENT_DETECT_EMAILS => '/content/detect_emails',
+            self::CONTENT_DETECT_SPAM => '/content/detect_spam',
+            self::CONTENT_SUMMARIZE => '/content/summarize',
+            self::CONTENT_KEYWORDS => '/content/keywords',
+            self::CONTENT_TRANSLATE => '/content/translate',
+            self::CONTENT_PARAPHRASE => '/content/paraphrase',
+            self::CONTENT_PROOFREAD => '/content/proofread',
+            self::SEO_GENERATE_TAGS => '/seo/generate_tags',
+        };
+    }
 
     public function label(): string
     {
@@ -49,6 +77,8 @@ enum SharpApiJobTypeEnum: string
             self::CONTENT_SUMMARIZE => 'Summarize Content',
             self::CONTENT_KEYWORDS => 'Generate Keywords/Tags',
             self::CONTENT_TRANSLATE => 'Translate Text',
+            self::CONTENT_PARAPHRASE => 'Paraphrase Text',
+            self::CONTENT_PROOFREAD => 'Proofread & Grammar check',
             self::SEO_GENERATE_TAGS => 'Generate SEO Tags',
         };
     }
@@ -71,6 +101,8 @@ enum SharpApiJobTypeEnum: string
             self::CONTENT_DETECT_EMAILS,
             self::CONTENT_DETECT_SPAM,
             self::CONTENT_TRANSLATE,
+            self::CONTENT_PARAPHRASE,
+            self::CONTENT_PROOFREAD,
             self::CONTENT_KEYWORDS,
             self::CONTENT_SUMMARIZE => 'Content',
             self::SEO_GENERATE_TAGS => 'SEO',